+++ app/components/budgets/budget_component.html.erb
@@ -33,13 +33,7 @@
       <% end %>
 
       <%= render Budgets::SupportsInfoComponent.new(budget) %>
-
-      <% unless budget.informing? %>
-        <div class="map inline">
-          <h2><%= t("budgets.index.map") %></h2>
-          <%= render_map(nil, "budgets", false, nil, coordinates) %>
-        </div>
-      <% end %>
+      <%= render Budgets::MapComponent.new(budget) %>
     </div>
   </div>
 </div>
