+++ app/components/budgets/map_component.html.erb
@@ -1,0 +1,4 @@
+<div class="map inline">
+  <h2><%= t("budgets.index.map") %></h2>
+  <%= render_map(nil, "budgets", false, nil, coordinates) %>
+</div>
